GH-592 Move log_* files only used by state_history into state_history library.
